requirements.txt
# PhishGard-AI/requirements.txt

# Core
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv         # Pour gérer les variables d'environnement (.env)

# LLM
transformers
torch

# OSINT
requests
python-whois
python-dateutil

# URL Analysis
pandas
numpy
scikit-learn
joblib
tldextract
python-whois
beautifulsoup4
selenium

# Web Framework
uvicorn 
fastapi